requirements.txt
ultralytics>=8.2.0
opencv-python>=4.8.0
numpy>=2.3.0
Pillow>=10.0.0
requests>=2.31.0
argostranslate>=1.9.0
piper-tts>=1.2.0

# PyTorch (will auto-detect CUDA if available)
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0